+++ changelog.md
@@ -1,5 +1,9 @@
 # The Modding Tree changelog:
 
+- Moved most of the code users will want to edit to mod.js.
+- Added getStartPoints()
+- Fixed issues with version number
+
 ### v2.0.5 - 10/16/20
 - Made more features (including prestige parameters) able to be dynamic.
 - Layer nodes can be hidden but still take up space with "ghost" visibility